Status service corrections

+++ gemsModules/complex/antibody/receive.py
@@ -1,4 +1,6 @@
 #!/usr/bin/env python3
+import os 
+
 from gemsModules.complex.antibody.json_string_manager import (
     Antibody_Json_String_Manager,
 )
@@ -23,10 +25,13 @@ def receive(incomingString: str) -> str:
     
     # Write the response to a a file in the project directory
     output_path = Antibody_manager.transaction.outputs.project.project_dir
-    response_json_path = output_path + "/response.json"
-    with open(response_json_path, "w") as f:
-        f.write(Antibody_manager.transaction.get_outgoing_string(prettyPrint=True))
-        f.write("\n")
+    if os.path.exists(output_path) is False:
+        log.debug("The output path does not exist")
+    else:
+        response_json_path = output_path + "/response.json"
+        with open(response_json_path, "w") as f:
+            f.write(Antibody_manager.transaction.get_outgoing_string(prettyPrint=True))
+            f.write("\n")
     
     # Return the response in a condensed format for the API
     return Antibody_manager.transaction.get_outgoing_string()
